requirements.txt
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=0.19.0
graphviz>=0.20.1
pyppeteer>=1.0.2
pillow==11.2.1